requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-telegram-bot[asyncio]
apscheduler
python-decouple
